Backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-storage-blob==12.22.0
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.6
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
isodate==0.7.2
numpy==1.26.4
packaging==24.2
pandas==2.2.0
psycopg2-binary==2.9.9
pyarrow==14.0.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.25
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
